backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
google-cloud-bigquery
google-cloud-firestore
firebase-admin
pandas-gbq